skill_priority_system/requirements.txt
google-generativeai
pandas
numpy
scikit-learn
python-dotenv